Avoid nested receives
The patch is a combined cherry-pick of commits:
0ded893693f380bc5b2c65f7d1e5b2c88f5309a5
069b22c5b24ff21d2e405fb7a443c5679c2a3687
adopted to the changes in the lager.

The main idea is to avoid nested receives. That's the reason
why the file is opened not in raw mode: it allows to send the
messages to IODevice and receive the response asynchronously.
Until the response is received, the messages are accumulated in the buffer.
If the buffer too big, the lager switches onto synchronous messaging mode.